Plot Energy

+++ src/datavis/shapes.rs
@@ -1,8 +1,8 @@
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
 use super::Vertex;
 
-pub fn create_cube(pos: [f64; 2], color: [f32; 4], s: f64) -> Vec<Vertex> {
+pub fn create_cube(pos: [f32; 2], color: [f32; 4], s: f32) -> Vec<Vertex> {
     let mut shape = Vec::with_capacity(6);
     shape.push(Vertex {
         position: [pos[0] - s, pos[1] - s],
@@ -35,27 +35,27 @@ pub fn create_cube(pos: [f64; 2], color: [f32; 4], s: f64) -> Vec<Vertex> {
     shape
 }
 
-pub fn create_circle(pos: [f64; 2], color: [f32; 4], r: f64, res: usize) -> Vec<Vertex> {
+pub fn create_circle(pos: [f32; 2], color: [f32; 4], r: f32, res: usize) -> Vec<Vertex> {
     let mut shape = Vec::with_capacity(3 * res);
-    let a = 2.0 * PI / res as f64;
+    let a = 2.0 * PI / res as f32;
 
     for i in 0..res {
-        let i = i as f64;
+        let i = i as f32;
         shape.push(Vertex {
             position: pos,
             color,
         });
         shape.push(Vertex {
             position: [
-                pos[0] + (r * f64::sin(a * i)),
-                pos[1] + (r * f64::cos(a * i)),
+                pos[0] + (r * f32::sin(a * i)),
+                pos[1] + (r * f32::cos(a * i)),
             ],
             color,
         });
         shape.push(Vertex {
             position: [
-                pos[0] + (r * f64::sin(a * (i + 1.0))),
-                pos[1] + (r * f64::cos(a * (i + 1.0))),
+                pos[0] + (r * f32::sin(a * (i + 1.0))),
+                pos[1] + (r * f32::cos(a * (i + 1.0))),
             ],
             color,
         });
@@ -64,7 +64,7 @@ pub fn create_circle(pos: [f64; 2], color: [f32; 4], r: f64, res: usize) -> Vec<
     shape
 }
 
-pub fn line(p1: [f64; 2], p2: [f64; 2], color: [f32; 4]) -> Vec<Vertex> {
+pub fn line(p1: [f32; 2], p2: [f32; 2], color: [f32; 4]) -> Vec<Vertex> {
     let mut shape = Vec::with_capacity(2);
 
     shape.push(Vertex {
